% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ann_run_example}
\alias{ann_run_example}
\title{Example data with run annotations}
\format{
## ann_run_example
A data frame with 6 rows and 3 columns:
\describe{
 \item{runID}{A column with run IDs in a form of F# (where # is a number). Each ID
 has to be unique.}
 \item{sampleID}{A column with sample IDs (one for each patient). It consists of
 a letter indicating a treatment group and a number indicating the patient
 number within the group. E.g. ID A2 stands for the second patient within a
 treatment group A. The same sample ID can be assigned to several run IDs.}
 \item{rep}{A column with repetition number. E.g. Run F4 is the second
 repetition for a patient A2.}

}
}
\source{
<https://raw.githubusercontent.com/TomChupan/data/main/data_example_run.csv>
}
\usage{
ann_run_example
}
\description{
This dataset contains run annotations for *data_example* dataset
to be used in a proteoME Shiny app.
}
\keyword{datasets}
