% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ann_sample_example}
\alias{ann_sample_example}
\title{Example data with sample annotations}
\format{
A data frame with 50 rows and 5 columns:
\describe{
 \item{sampleID: }{A column with **unique** sample IDs (one for each patient). It consists of
 a letter indicating a treatment group and a number indicating the patient
 number within the group. E.g. ID A2 stands for the second patient within a
 treatment group A.}
 \item{treatment: }{A column with a treatment group specification.}
 \item{sex: }{Male (M) or female (F).}
 \item{age: }{Age in years.}
 \item{weight_kg: }{Weight in kg.}


}
}
\source{
<https://raw.githubusercontent.com/TomChupan/data/main/data_example_sample2.csv>
}
\usage{
ann_sample_example
}
\description{
This dataset contains sample annotations for *data_example* dataset
to be used in a proteoME Shiny app.
}
\keyword{datasets}
