% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{Example data with protein abundances}
\format{
A data frame with 2018 rows and 101 columns:
\describe{
\itemize{
 \item{Accession: }{An UniProt accession number – a unique identifier assigned
 to a protein entry in the UniProt Knowledgebase (UniProtKB).
 See \url{https://www.uniprot.org/} }
 \item{F#: }{Columns with protein abundances for each run
 (several runs may come from one patient – replicates)}
 \itemize{
 \item Abundances can be raw, imputed or normalized (or both imputed and
  normalized)
  }
 }

}
}
\source{
<https://raw.githubusercontent.com/TomChupan/data/main/data_example3.csv>
}
\usage{
data_example
}
\description{
This dataset contains sample data with protein abundances to be used in a
proteoME Shiny app.
}
\keyword{datasets}
