% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aggregate.R
\name{mod_aggregate_ui}
\alias{mod_aggregate_ui}
\alias{mod_aggregate_server}
\title{#' @title mod_aggregate_ui and mod_aggregate_server}
\usage{
mod_aggregate_ui(id)

mod_aggregate_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A 'reactiveValues()' list containing (among other objects) data set with
abundances. This is aggregated here with one of the available methods and
parameters to the sample level. A logical indicator (aggregated or not?) is
toggled here. Aggregation method and parameters are saved to r to be used in
other modules.}
}
\description{
A shiny Module containing sidebar content for data aggregation.
}
