% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis.R
\name{mod_analysis_ui}
\alias{mod_analysis_ui}
\alias{mod_analysis_server}
\title{#' @title mod_analysis_ui and mod_analysis_server}
\usage{
mod_analysis_ui(id)

mod_analysis_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A 'reactiveValues()' list containing (among other objects) data set with
abundances. This is analysed here with one of the available methods.
A logical indicator (analysed or not?) is toggled here.}
}
\description{
A shiny Module containing sidebar content for data analysis
}
