% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_body.R
\name{mod_body_ui}
\alias{mod_body_ui}
\alias{mod_body_server}
\title{mod_body_ui and mod_body_server}
\usage{
mod_body_ui(id)

mod_body_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A "storage" for the variables used throughout the app}
}
\description{
A shiny Module containing tha main body content. This module is
included in a dashboardBody() function within app_ui.R.
The body of the app is divided into __ tabs:
\itemize{
\item{Data import}{Tab for uploading 3 types of data (protein abundances,
run annotations, sample annotations) and their basic specification}
}
}
\details{
body UI Function
}
