% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_body_analysis.R
\name{mod_body_analysis_ui}
\alias{mod_body_analysis_ui}
\alias{mod_body_analysis_server}
\title{mod_body_analysis_ui and mod_body_analysis_server}
\usage{
mod_body_analysis_ui(id, box_title = "Your title.")

mod_body_analysis_server(id, validate_message, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{box_title}{Character to be used as a box title.}

\item{validate_message}{Message to be displayed when there are no data to build a table.}

\item{r}{A 'reactiveValues()' list containing ... (will be specified)}
}
\description{
A shiny module for showing result table and volcano plot in the app body.
}
