% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_body_impute.R
\name{mod_body_impute_ui}
\alias{mod_body_impute_ui}
\alias{mod_body_impute_server}
\title{mod_body_impute_ui and mod_body_impute_server}
\usage{
mod_body_impute_ui(id, box_title = "Your title.")

mod_body_impute_server(id, validate_message, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{box_title}{Character to be used as a box title.}

\item{r}{A 'reactiveValues()' list containing (among other objects) ___.
These are used here in a ___ function.}
}
\description{
A shiny Module for creating content on the imputation tab
in the app body.
}
