% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_body_normalize.R
\name{mod_body_normalize_ui}
\alias{mod_body_normalize_ui}
\alias{mod_body_normalize_server}
\title{mod_body_normalize_ui and mod_body_normalize_server}
\usage{
mod_body_normalize_ui(id, box_title)

mod_body_normalize_server(id, validate_message, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{box_title}{Character to be used as a box title.}

\item{validate_message}{Message to be displayed when there are no data to build a boxplot.}

\item{r}{A 'reactiveValues()' list containing (among other objects) all three
datasets and also a long-format version (combination of these datasets).
This is used in the boxplot rendering. We also use here the selected normalization
method from the mod_normalize module, the active tab and the information whether
the dataset was already normalized (r$normalizedTF).}
}
\description{
A shiny module for creating boxplots (static and reactive)
in the app body to help decide whether/how to normalize data.
}
