% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_body_plot.R
\name{mod_body_plot_ui}
\alias{mod_body_plot_ui}
\alias{mod_body_plot_server}
\title{mod_body_plot_ui and mod_body_plot_server}
\usage{
mod_body_plot_ui(id, box_title = "Your title.")

mod_body_plot_server(id, plot_type, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{box_title}{Character to be used as a box title.}

\item{plot_type}{Character to specify the plot. It should be made up of 3 parts
separated by an underscore (_): \strong{tab ID}, \strong{plot type} and
\strong{order of the plot of the same type on the same tab}.
For example: eda_box_1 stands for the first box plot on the "Exploratory data analysis"
tab.}

\item{r}{A 'reactiveValues()' list containing (among other objects) plots built
in mod_plot. These are used here in a renderGirafe() function.}
}
\description{
A shiny Module for creating plots in the app body.
}
