% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_body_sumtab.R
\name{mod_body_sumtab_ui}
\alias{mod_body_sumtab_ui}
\alias{mod_body_sumtab_server}
\title{mod_body_sumtab_ui and mod_body_sumtab_server}
\usage{
mod_body_sumtab_ui(id)

mod_body_sumtab_server(id, box_title = "Your title.", validate_message, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{box_title}{Character to be used as a box title.}

\item{validate_message}{Message to be displayed when there are no data to build a table.}

\item{r}{A 'reactiveValues()' list containing (among other objects) all three
datasets and also a long-format version (combination of these datasets).
This is used in the summary tables. We also use here the selected transformation
method from the mod_transform module, the active tab and the information whether
the dataset was already transformed (r$transformedTF).}
}
\description{
A shiny for creating summary tables in the app body.
}
