% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_body_tab1.R
\name{mod_body_tab1_ui}
\alias{mod_body_tab1_ui}
\alias{mod_body_tab1_server}
\title{mod_body_tab1_ui and mod_body_tab1_server}
\usage{
mod_body_tab1_ui(id, box_title = "Your title.", data_type = c(1, 2, 3, 4))

mod_body_tab1_server(
  id,
  validate_message = "Please upload missing file!",
  data_type = c(1, 2, 3, 4),
  r
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{box_title}{Character to be used as a box title (there is one of the 3
possible tables in that box).}

\item{data_type}{Number from 1 to 3 defining the type of the data
as following:
\itemize{
 \item 1: "Abundances"
 \item 2: "Annotations (run)"
 \item 3: "Annotations (sample)"
}}

\item{validate_message}{Message to be displayed when there are no data to build a table.}

\item{r}{A "storage" for the variables used throughout the app}
}
\description{
A shiny Module with the main body content connected with data import.
}
\details{
body_tab1 UI Function
}
