% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_factors.R
\name{mod_factors_ui}
\alias{mod_factors_ui}
\alias{mod_factors_server}
\title{mod_factors_ui and mod_factors_server}
\usage{
mod_factors_ui(id)

mod_factors_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A "storage" for the variables used throughout the app}
}
\description{
A shiny Module for setting factor variables and editing factor levels
}
\details{
factors UI Function
}
