% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filter_ui}
\alias{mod_filter_ui}
\alias{mod_filter_server}
\title{#' @title mod_filter_ui and mod_filter_server}
\usage{
mod_filter_ui(id)

mod_filter_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A 'reactiveValues()' list containing (among other objects) data set with
abundances. This is filtered here with one of the available methods and
parameters (and it can be downloaded).
A logical indicator (filtered or not?) is toggled here.
Filtering method and parameters are saved to r to be used in other modules.}
}
\description{
A shiny Module containing sidebar content for data filtering.
}
