% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_impute.R
\name{mod_impute_ui}
\alias{mod_impute_ui}
\alias{mod_impute_server}
\title{#' @title mod_impute_ui and mod_impute_server}
\usage{
mod_impute_ui(id)

mod_impute_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A 'reactiveValues()' list containing (among other objects) data set with
abundances. This is imputed here with one of the available methods and
parameters (and it can be downloaded).
A logical indicator (imputed or not?) is toggled here.
Imputation method and parameters are saved to r to be used in other modules.}
}
\description{
A shiny Module containing sidebar content for data imputation
}
