% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normalize.R
\name{mod_normalize_ui}
\alias{mod_normalize_ui}
\alias{mod_normalize_server}
\title{mod_normalize_ui and mod_normalize_server}
\usage{
mod_normalize_ui(id)

mod_normalize_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A 'reactiveValues()' list containing (among other objects) data set with
abundances. This is normalized here with one of the available methods. The original
data set is overwritten. A logical indicator (normalized or not?) is
toggled here. Normalization method is saved to r to be used in
other modules.}
}
\description{
A shiny Module containing sidebar content for data normalization.
}
