% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_ui}
\alias{mod_plot_ui}
\alias{mod_plot_server}
\title{mod_plot_ui and mod_plot_server}
\usage{
mod_plot_ui(id, plot_type, menuItem_label)

mod_plot_server(id, plot_type, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{plot_type}{Character to specify the plot. It should be made up of 3 parts
separated by an underscore (_):
\itemize{
 \item Tab ID
 \item Plot type
 \item Order of the plot of the same type on the same tab
}
For example: eda_box_1 stands for the first box plot on the "Exploratory data analysis"
tab.}

\item{menuItem_label}{Character to name the menu item. It should be the name
and a short description of the plot (e.g. 'Boxplot of abundances by runs').}

\item{r}{A "storage" for the variables used throughout the app}
}
\description{
A shiny Module containing sidebar content for making plots.
}
