% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_transform.R
\name{mod_transform_ui}
\alias{mod_transform_ui}
\alias{mod_transform_server}
\title{mod_transform_ui and mod_transform_server}
\usage{
mod_transform_ui(id)

mod_transform_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A 'reactiveValues()' list containing (among other objects) data set with
abundances. This is transformed here with one of the available methods. The original
data set is overwritten.}
}
\description{
A shiny Module containing sidebar content for data transformation.
}
