% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_ui}
\alias{mod_upload_ui}
\alias{mod_upload_server}
\title{mod_upload_ui and mod_upload_server}
\usage{
mod_upload_ui(
  id,
  menuItem_label = c("Abundances", "Annotations (run)", "Annotations (sample)"),
  fileInput_label
)

mod_upload_server(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{menuItem_label}{Character to name the menu item. For upload menu,
it can take one of the following values:
\itemize{
 \item "Abundances"
 \item "Annotations (run)"
 \item "Annotations (sample)"
}}

\item{fileInput_label}{Character to be used as a fileInput label.}
}
\description{
A shiny Module for data upload.
}
\details{
upload UI Function
}
