% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_proteoFI.R
\name{proteoFI}
\alias{proteoFI}
\title{proteoAG}
\usage{
proteoFI(
  data_aggreg,
  ann_sample,
  n = 30,
  method = c("all", "onegroup", "eachgroup")
)
}
\arguments{
\item{data_aggreg}{Data frame with the aggregated abundances on the sample level. First column
contains Accession numbers, other columns (named with sampleID's) contains protein abundances of
every sample/column.}

\item{ann_sample}{Data frame with sample annotations in the same form as
ann_sample_example (see ?proteoME::ann_sample_example).}

\item{method}{It can be one of these 3 values: 1) "all": protein is kept in the dataset
when it is quantified in at least 'percent' % of all samples. 2) "onegroup":
protein is kept in the dataset when it is quantified in at least 'percent' %
of samples within at least one treatment group. 3) "eachgroup":
protein is kept in the dataset when it is quantified in at least 'percent' %
of samples within each treatment group.}

\item{percent}{A number between 1 and 100 that represents the minimum required percentage
of samples with a quantified protein. Every protein with relative quantification
below this value (depending on the selected method - see 'method') will be removed.}
}
\value{
A data frame with the filtered abundances.
}
\description{
A function for filtering the abundances data (its rows with proteins)
based on NA's.
}
\examples{
d_filtered=proteoFI(data_example,ann_sample_example,percent=30,method="onegroup")

}
