% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsML.R
\name{run_plsda}
\alias{run_plsda}
\title{PLS-DA of a proteomics matrix}
\usage{
run_plsda(expr, groups, ncomp = 2)
}
\arguments{
\item{expr}{Numeric matrix of expression values, proteins x samples.}

\item{groups}{Factor of group labels, length \code{ncol(expr)}.}

\item{ncomp}{Number of PLS components to fit. Default \code{2}.}
}
\value{
A list with components:
\item{scores}{Data frame of Comp1/Comp2 scores per sample.}
\item{plot}{ggplot scatter plot of the first two components.}
\item{model}{The full \code{plsr} model object.}
}
\description{
Runs partial least squares discriminant analysis via
\code{pls::plsr()} with a one-hot encoded group response. Requires the
\pkg{pls} package.
}
\examples{
\dontrun{
csv_path <- system.file("extdata", "example_proteome.csv", package = "ProteomicsML")
io <- read_wide_proteomics(csv_path)
pls_res <- run_plsda(io$expr, io$groups)
pls_res$plot
}
}
