% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ann_run_example}
\alias{ann_run_example}
\title{Example data with run annotations}
\format{
A data frame with 100 rows and 6 columns:
\describe{
 \item{runID}{A column with run IDs in a form of F# (where # is a number). Each ID
 has to be unique.}
 \item{sampleID}{A column with sample IDs (one for each patient). It consists of
 a letter indicating a treatment group and a number indicating the patient
 number within the group. E.g. ID A2 stands for the second patient within a
 treatment group A. The same sample ID can be assigned to several run IDs.}
 \item{rep}{A column with replicate number. E.g. Run F4 is the second
 replicate for a patient A2.}
 \item{batchNo}{Batch number.}
 \item{batchOrder}{Order of the run within the batch.}
 \item{dateTime}{Date and time of the measurement in a \%m/\%d/\%Y \%R format.}

}
}
\source{
<https://raw.githubusercontent.com/TomChupan/data/main/data_example_run3.csv>
}
\usage{
ann_run_example
}
\description{
This dataset contains run annotations for \strong{data_example} dataset
to be used in a proteoME Shiny app.
}
\keyword{datasets}
