% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{mod_sidebar_ui}
\alias{mod_sidebar_ui}
\alias{mod_sidebar_server}
\title{mod_sidebar_ui and mod_sidebar_server}
\usage{
mod_sidebar_ui(id)

mod_sidebar_server(id, r)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{r}{A "storage" for the variables used throughout the app}
}
\description{
A shiny Module containing sidebar content
}
\details{
sidebar UI Function
}
